massive rework read description
move radar out of App.js
Main.js works as a router
Pages in /pages
radar moved into RadarPage
Statistics created then moved into StatisticsPage
Projects moved to ProjectsPage and then the modal created into component to be used on Radar and Projects page
Header updated to use menudropdown
menudropdown created
HomePage created to show the 3 current sections
respective CSS files added
react-router-dom dependecy added for navigation
minor UI refinements
