+++ frontend/package.json
@@ -8,16 +8,19 @@
     "@aws-sdk/credential-provider-node": "^3.699.0",
     "@aws-sdk/credential-providers": "^3.699.0",
     "@aws-sdk/s3-request-presigner": "^3.705.0",
+    "@radix-ui/react-popover": "^1.1.4",
     "@testing-library/jest-dom": "^5.17.0",
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
+    "date-fns": "^4.1.0",
     "http-proxy-middleware": "^2.0.6",
     "papaparse": "^5.4.1",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
     "react-hot-toast": "^2.4.1",
     "react-icons": "^5.3.0",
-    "react-scripts": "5.0.1",
+    "react-router-dom": "^7.1.1",
+    "react-scripts": "^5.0.1",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
@@ -53,5 +56,8 @@
     "eslint-plugin-react": "^7.34.0",
     "eslint-plugin-react-hooks": "^4.6.0",
     "prettier": "^3.2.5"
+  },
+  "resolutions": {
+    "nth-check": "^2.0.1"
   }
 }
