+++ frontend/src/components/Header/Header.js
@@ -1,6 +1,6 @@
 import React from "react";
-import ThemeToggle from "../ThemeToggle/ThemeToggle";
-// import FileUpload from "./FileUpload";
+import { useNavigate } from "react-router-dom";
+import MenuDropdown from "../MenuDropdown/MenuDropdown";
 import "../../styles/Header.css";
 import Logo from "../../assets/logo.png";
 import { IoClose, IoSearch } from "react-icons/io5";
@@ -19,6 +19,8 @@ import { IoClose, IoSearch } from "react-icons/io5";
  * @param {Function} props.onFileUpload - Function to call when a file is uploaded.
  * @param {Function} props.checkForDuplicates - Function to call to check for duplicates.
  * @param {Function} props.onOpenProjects - Function to call when the projects button is clicked.
+ * @param {Function} props.onStatsTechClick - Function to call when a technology is clicked.
+ * @param {boolean} props.hideSearch - Whether to hide the search bar.
  */
 function Header({
   searchTerm,
@@ -26,66 +28,72 @@ function Header({
   searchResults,
   onSearchResultClick,
   onOpenProjects,
+  onStatsTechClick,
+  hideSearch = false
 }) {
+  const navigate = useNavigate();
+
   /**
    * Clears the search term.
    */
   const clearSearch = () => {
     onSearchChange("");
   };
+
   return (
     <header className="radar-header">
       <div className="header-left">
-        <img src={Logo} alt="Logo" className="logo" />
-        <h1>Tech Radar</h1>
+        <img 
+          src={Logo} 
+          alt="Logo" 
+          className="logo" 
+          onClick={() => navigate('/radar')}
+          style={{ cursor: 'pointer' }}
+        />
+        <h1 onClick={() => navigate('/')} style={{ cursor: 'pointer' }}>Digital Landscape</h1>
       </div>
       <div className="header-right">
-        <div className="search-container">
-          <IoSearch className="search-icon" />
-
-          <input
-            type="text"
-            placeholder="Search technologies..."
-            value={searchTerm}
-            onChange={(e) => onSearchChange(e.target.value)}
-            className="search-input"
-          />
-          <button
-            className="search-clear"
-            onClick={clearSearch}
-            style={{ display: searchTerm ? "block" : "none" }}
-          >
-            {" "}
-            <IoClose />{" "}
-          </button>
-          {searchResults.length > 0 && (
-            <div className="search-results">
-              {searchResults.map((result) => (
-                <div
-                  key={result.id}
-                  className="search-result-item"
-                  onClick={() => onSearchResultClick(result)}
-                >
-                  <span className="search-result-title">{result.title}</span>
-                  <span
-                    className={`search-result-ring ${result.timeline[0].ringId.toLowerCase()}`}
+        {!hideSearch && (
+          <div className="search-container">
+            <IoSearch className="search-icon" />
+            <input
+              type="text"
+              placeholder="Search technologies..."
+              value={searchTerm}
+              onChange={(e) => onSearchChange(e.target.value)}
+              className="search-input"
+            />
+            <button
+              className="search-clear"
+              onClick={clearSearch}
+              style={{ display: searchTerm ? "block" : "none" }}
+            >
+              <IoClose />
+            </button>
+            {searchResults.length > 0 && (
+              <div className="search-results">
+                {searchResults.map((result) => (
+                  <div
+                    key={result.id}
+                    className="search-result-item"
+                    onClick={() => onSearchResultClick(result)}
                   >
-                    {result.timeline[0].ringId}
-                  </span>
-                </div>
-              ))}
-            </div>
-          )}
-        </div>
-
-        <button 
-          className="projects-button"
-          onClick={onOpenProjects}
-        >
-          View Projects
-        </button>
-
-        <ThemeToggle />
+                    <span className="search-result-title">{result.title}</span>
+                    <span
+                      className={`search-result-ring ${result.timeline[0].ringId.toLowerCase()}`}
+                    >
+                      {result.timeline[0].ringId}
+                    </span>
+                  </div>
+                ))}
+              </div>
+            )}
+          </div>
+        )}
+        <MenuDropdown 
+          onOpenProjects={onOpenProjects} 
+          onStatsTechClick={onStatsTechClick}
+        />
       </div>
     </header>
   );
