+++ frontend/src/components/ThemeToggle/ThemeToggle.js
@@ -21,7 +21,9 @@ function ThemeToggle() {
       className="theme-toggle"
       aria-label={`Switch to ${theme === "light" ? "dark" : "light"} mode`}
     >
+      
       {theme === "light" ? <SunIcon /> : <MoonIcon />}
+      Toggle Theme
     </button>
   );
 }
