+++ frontend/src/styles/HomePage.css
@@ -1,0 +1,106 @@
+.home-page {
+  width: 100%;
+  min-height: calc(100vh - 70px);
+  margin-top: 70px;
+  background: hsl(var(--background));
+}
+
+.home-content {
+  width: 100%;
+  text-align: left;
+  padding: 32px;
+  box-sizing: border-box;
+  display: flex;
+  flex-direction: column;
+  align-items: flex-start;
+  gap: 32px;
+}
+
+.home-content-header {
+  width: 100%;
+  text-align: center;
+}
+
+.home-content h1 {
+  font-size: 32px;
+  width: 100%;
+  text-align: center;
+  margin: 0;
+}
+
+.home-content-header > p {
+  font-size: 16px;
+  color: hsl(var(--muted-foreground));
+  margin: 0;
+  margin-top: 8px;
+}
+
+.navigation-cards {
+  display: grid;
+  grid-template-columns: repeat(auto-fit, minmax(300px, 1fr));
+  gap: 24px;
+  /* border: 1px solid hsl(var(--border)); */
+  border-radius: var(--radius);
+  padding: 16px;
+  box-sizing: border-box;
+  width: 100%;
+}
+
+.nav-card {
+  background: hsl(var(--card));
+  border: 1px solid hsl(var(--border));
+  border-radius: var(--radius);
+  padding: 16px;
+  cursor: pointer;
+  transition: all 0.3s ease;
+  display: flex;
+  flex-direction: column;
+  align-items: flex-start;
+  text-align: left;
+  gap: 8px;
+}
+
+.nav-card:hover {
+  box-shadow: 0 8px 24px hsl(var(--foreground) / 0.1);
+  border-color: hsl(var(--primary));
+}
+
+.nav-card svg {
+  color: hsl(var(--primary));
+  width: 24px;
+  height: 24px;
+  margin-bottom: 2px;
+}
+
+.nav-card h2 {
+  font-size: 18px;
+  margin: 0;
+}
+
+.nav-card p {
+  color: hsl(var(--muted-foreground));
+  margin: 0;
+  /* line-height: 1.6; */
+  font-size: 14px;
+}
+
+.nav-card-header {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  gap: 8px;
+}
+
+@media (max-width: 768px) {
+  .home-content h1 {
+    font-size: 36px;
+  }
+
+  .home-content > p {
+    font-size: 18px;
+  }
+
+  .navigation-cards {
+    grid-template-columns: 1fr;
+  }
+}
