+++ frontend/src/styles/ProjectModal.css
@@ -1,0 +1,167 @@
+.modal-overlay {
+  position: fixed;
+  top: 0;
+  left: 0;
+  right: 0;
+  bottom: 0;
+  background: hsl(var(--background) / 0.8);
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  z-index: 1000;
+  backdrop-filter: blur(4px);
+}
+
+.modal-content {
+  background: hsl(var(--card));
+  border-radius: var(--radius);
+  padding: 24px;
+  max-width: 800px;
+  width: 90%;
+  max-height: 90vh;
+  overflow-y: auto;
+  position: relative;
+  border: 1px solid hsl(var(--border));
+  box-shadow: 0 4px 10px 0 hsl(var(--foreground) / 0.1);
+}
+
+.modal-close {
+  background: none;
+  border: none;
+  color: hsl(var(--foreground));
+  font-size: 24px;
+  cursor: pointer;
+}
+
+.project-header {
+  margin-bottom: 24px;
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+}
+
+.project-header-left {
+  display: flex;
+  flex-direction: column;
+  align-items: flex-start;
+}
+
+.project-header-right {
+  display: flex;
+  align-items: flex-start;
+}
+
+.project-title-section h2 {
+  margin: 0;
+  font-size: 24px;
+  font-weight: bold;
+}
+
+.project-short-name {
+  color: hsl(var(--muted-foreground));
+  font-size: 14px;
+  margin-top: 4px;
+}
+
+
+.project-item {
+    cursor: pointer;
+    padding: 8px;
+    margin: 4px 0;
+    border-radius: 4px;
+    background-color: hsl(var(--muted));
+    transition: background-color 0.2s ease;
+  }
+  
+  .project-item:hover {
+    background-color: hsl(var(--muted) / 0.8);
+  }
+  
+  .project-modal {
+    max-width: 800px;
+    background-color: hsl(var(--card));
+    border: 1px solid hsl(var(--border));
+  }
+  
+  .detail-item {
+    background-color: hsl(var(--muted));
+    padding: 15px;
+    border-radius: 8px;
+  }
+  
+  .detail-item h3 {
+    margin: 0 0 8px 0;
+    font-size: 1em;
+    color: hsl(var(--foreground));
+  }
+  
+  .detail-item a {
+    color: #0069e5;
+    text-decoration: none;
+    font-size: 14px;
+  }
+  
+
+
+@media (max-width: 768px) {
+  .modal-content {
+    width: 95%;
+    padding: 16px;
+  }
+
+  .project-title-section h2 {
+    font-size: 20px;
+  }
+
+  .detail-item h3,
+  .detail-item p {
+    font-size: 12px;
+  }
+} 
+
+
+.project-details {
+    display: grid;
+    grid-template-columns: repeat(auto-fill, minmax(300px, 1fr));
+    gap: 16px;
+    margin-top: 20px;
+  }
+  
+  .detail-item {
+    background-color: hsl(var(--muted));
+    padding: 16px;
+    border-radius: var(--radius);
+    height: fit-content;
+    text-align: left;
+  }
+  
+  .detail-item h3 {
+    margin: 0 0 8px 0;
+    font-size: 14px;
+    color: hsl(var(--muted-foreground));
+  }
+  
+  .detail-item p {
+    margin: 0;
+    font-size: 14px;
+    line-height: 1.5;
+    color: hsl(var(--foreground));
+  }
+  
+  .detail-item a:hover {
+    text-decoration: underline;
+  }
+  
+  .project-modal {
+    width: 50vw;
+    max-width: 1200px;
+    max-height: 50vh;
+    overflow-y: auto;
+  }
+  
+  @media (max-width: 1400px) {
+    .project-modal {
+      width: 90vw;
+      max-height: 90vh;
+    }
+  }
