+++ frontend/src/styles/Projects.css
@@ -1,31 +1,19 @@
-.projects-modal-overlay {
-  position: fixed;
-  top: 0;
-  left: 0;
-  right: 0;
-  bottom: 0;
-  background-color: hsl(var(--background) / 0.8);
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  z-index: 1000;
-}
+
 
 .projects-modal-content {
   background: hsl(var(--card));
-  border-radius: var(--radius);
-  padding: 16px 16px 16px 16px;
   box-sizing: border-box;
-  width: 90%;
-  max-width: 1200px;
-
+  width: 100%;
+  text-align: left;
 
   position: relative;
-  border: 1px solid hsl(var(--border));
   scrollbar-width: 1px;
   scrollbar-color: hsl(var(--border)) hsl(var(--card));
 }
 
+.projects-modal-content h2 {
+  text-align: left;
+}
 .projects-modal-content::-webkit-scrollbar {
   width: 1px;
 }
@@ -47,7 +35,7 @@
 
 .projects-modal-content h2 {
   margin: 0;
-  font-size: 20px;
+  font-size: 24px;
   color: hsl(var(--foreground));
 }
 
@@ -61,7 +49,7 @@ font-size: 14px;
   flex-direction: column;
   overflow-y: auto;
   overflow-x: hidden;
-  height: 50vh;
+  height: 70vh;
   gap: 6px;
   padding: 16px 0;
   margin-top: 0;
@@ -99,6 +87,7 @@ font-size: 14px;
   text-decoration: underline;
   text-decoration-style: dotted;
   text-underline-offset: 4px;
+  font-size: 14px;
 }
 
 .project-name-short {
