+++ frontend/src/styles/ThemeToggle.css
@@ -1,19 +1,16 @@
 .theme-toggle {
-  background: hsl(var(--background));
-  border: 1px solid hsl(var(--border));
-  cursor: pointer;
-  font-size: 20px;
-  border-radius: var(--radius);
-  transition: all 0.2s ease;
-  box-sizing: border-box;
-  padding: 10px 15px;
-  font-size: 14px;
+  width: 100%;
+  padding: 12px 16px;
+  text-align: left;
+  background: none;
+  border: none;
   color: hsl(var(--foreground));
-  height: 33px;
+  cursor: pointer;
   display: flex;
+  border-radius: 0 0 var(--radius) var(--radius);
   align-items: center;
-  justify-content: center;
   gap: 8px;
+  transition: all 0.2s ease;
 }
 
 .theme-toggle:hover {
